using let, moving enum class to another file, testing negative scenarios and using singular verbs in it blocks
