+++ main.rb
@@ -14,3 +14,5 @@ def multiply_els(arr)
 end
 
 puts multiply_els([2, 4, 5])
+
+puts ["one", "", "true"].my_any?(/f/)
